update ignores and move types
