+++ .gitignore
@@ -5,6 +5,7 @@
 config.guess
 config.sub
 config.log
+config.status
 Makefile.in
 aclocal.m4
 compile
@@ -13,5 +14,95 @@ depcomp
 install-sh
 ltmain.sh
 missing
+Makefile
+error.log
 *~
 test-driver
+/m4
+/m4/
+/m4/*
+/m4/pkg.m4
+/m4/sdl.m4
+/bflibrary/m4/pkg.m4
+/bflibrary/m4/sdl.m4
+/bflibrary/m4
+/bflibrary/m4/
+/bflibrary/m4/*
+/bflibrary/.deps/*
+/bflibrary/src/general/.deps/*
+/bflibrary/src/general/.deps/.dirstamp
+/bflibrary/src/general/.dirstamp
+/bflibrary/Makefile
+/bflibrary/referenc/*.png
+/bflibrary/config.status
+/bflibrary/include/stamp-h1
+/bflibrary/libbullfrog.*
+/bflibrary/src/general/*.o
+/bflibrary/src/x86-win-sdl/.deps/.dirstamp
+/bflibrary/src/x86-win-sdl/.dirstamp
+/bflibrary/src/x86-win-sdl/.deps/*
+/bflibrary/src/x86-win-sdl/*.o
+/bflibrary/tests/.deps/*
+/bflibrary/tests/.deps/.dirstamp
+/bflibrary/tests/.dirstamp
+/src/.deps/*
+/src/Makefile
+/src/*.o
+/src/swars
+/src/stamp-h1
+/src/error.log
+/src/config.h
+/src/wrappers_dos.sx
+/src/wrappers_game.sx
+/src/wrappers_libc.sx
+/src/wrappers_util.sx
+/util/Makefile
+/res/swars-config.nsh
+/lang/eng/alltext-p.idx
+/lang/eng/alltext-p.wad
+/lang/eng/alltext-s.idx
+/lang/eng/alltext-s.wad
+/lang/eng/alltext-z.idx
+/lang/eng/alltext-z.wad
+/lang/fre/alltext-p.idx
+/lang/fre/alltext-p.wad
+/lang/fre/alltext-s.idx
+/lang/fre/alltext-s.wad
+/lang/fre/alltext-z.idx
+/lang/fre/alltext-z.wad
+/lang/ger/alltext-p.idx
+/lang/ger/alltext-p.wad
+/lang/ger/alltext-s.idx
+/lang/ger/alltext-s.wad
+/lang/ger/alltext-z.idx
+/lang/ger/alltext-z.wad
+/lang/ita/alltext-p.idx
+/lang/ita/alltext-p.wad
+/lang/ita/alltext-s.idx
+/lang/ita/alltext-s.wad
+/lang/ita/alltext-z.idx
+/lang/ita/alltext-z.wad
+/lang/spa/alltext-p.idx
+/lang/spa/alltext-p.wad
+/lang/spa/alltext-s.idx
+/lang/spa/alltext-s.wad
+/lang/spa/alltext-z.idx
+/lang/spa/alltext-z.wad
+/lang/swe/alltext-p.idx
+/lang/swe/alltext-p.wad
+/lang/swe/alltext-s.idx
+/lang/swe/alltext-s.wad
+/lang/swe/alltext-z.idx
+/lang/swe/alltext-z.wad
+/bflibrary/CMakeFiles
+/bflibrary/cmake_install.cmake
+/bflibrary/install_manifest.txt
+/bflibrary/CMakeCache.txt
+/CMakeFiles
+cmake_install.cmake
+install_manifest.txt
+CMakeCache.txt
+/bfsoundlib/CMakeFiles
+/bfsoundlib/cmake_install.cmake
+/bfsoundlib/install_manifest.txt
+/bfsoundlib/CMakeCache.txt
