+++ bflibrary/.gitignore
@@ -11,3 +11,32 @@ install-sh
 ltmain.sh
 missing
 *~
+/m4/pkg.m4
+/m4/sdl.m4
+/.deps/*
+/src/config.h
+/src/general/*.o
+/src/general/.deps/*
+/src/general/.deps/.dirstamp
+/src/general/.dirstamp
+/src/x86-win-sdl/.deps/.dirstamp
+/src/x86-win-sdl/.dirstamp
+/src/x86-win-sdl/.deps/*
+/src/x86-win-sdl/*.o
+/tests/.deps/*
+/tests/.deps/.dirstamp
+/tests/.dirstamp
+/m4
+/m4/
+/m4/*
+/m4/pkg.m4
+/m4/sdl.m4
+Makefile
+/referenc/*.png
+config.status
+/include/stamp-h1
+libbullfrog.*
+/CMakeFiles
+cmake_install.cmake
+install_manifest.txt
+CMakeCache.txt
